fixed total count

+++ src/services/proposal.rs
@@ -5,14 +5,14 @@ impl Services {
     pub async fn search_proposals(
         &self,
         input: SearchProposalsRequest,
-    ) -> Result<(Vec<ProposalFromDb>, i32), anyhow::Error> {
+    ) -> Result<(Vec<ProposalFromDb>, i64), anyhow::Error> {
         self.sqlx_client.search_proposals(input).await
     }
 
     pub async fn search_votes(
         &self,
         input: SearchVotesRequest,
-    ) -> Result<(Vec<VoteFromDb>, i32), anyhow::Error> {
+    ) -> Result<(Vec<VoteFromDb>, i64), anyhow::Error> {
         self.sqlx_client.search_votes(input).await
     }
 }
